Update README.md
The VBA Homework Stock Data Analysis Script allows to go through the yearly changes, the price percent change for the year, and the total volume traded for that year. It will then sort through the calculated data and determine the greatest percentage increase, the greatest percentage decrease, and the stock with the greatest volume of the year. The VBA script also allows the user to  cycle through each sheet automatically.
From the practical perspectives the script was first run through the Test Data, after that through real life Stock Data.

The color formatting in Developer mode was  done as well to visualise the yearly change and yearly percentage change required in the main part of VBA Stock Data Homework.